backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-limiter==3.9.4
psycopg2-binary==2.9.10
openai==1.78.0
chromadb==0.6.3
pypdf==5.4.0
python-docx==1.1.2
pytest==8.3.5
requests==2.32.3
werkzeug==3.1.3
